+++ skia-safe/src/gpu/direct_context.rs
@@ -102,12 +102,10 @@ impl DirectContext {
 
     #[cfg(feature = "metal")]
     pub fn new_metal<'a>(
-        backend: &crate::gpu::mtl::BackendContext,
+        backend_context: &crate::gpu::mtl::BackendContext,
         options: impl Into<Option<&'a ContextOptions>>,
     ) -> Option<DirectContext> {
-        DirectContext::from_ptr(unsafe {
-            sb::C_GrContext_MakeMetal(backend.native(), options.into().native_ptr_or_null())
-        })
+        crate::gpu::direct_contexts::make_metal(backend_context, options)
     }
 
     #[cfg(feature = "d3d")]
