+++ skia-safe/src/gpu/ganesh/gl.rs
@@ -1,5 +1,9 @@
 mod backend_surface;
 mod direct_context;
+#[cfg(target_arch = "wasm32")]
+mod make_web_gl_interface;
 
 pub use backend_surface::*;
 pub use direct_context::*;
+#[cfg(target_arch = "wasm32")]
+pub use make_web_gl_interface::*;
