+++ skia-safe/src/gpu/ganesh/mtl.rs
@@ -1,3 +1,9 @@
+mod backend_context;
+mod direct_context;
 mod surface_metal;
+pub(crate) mod types;
 
+pub use backend_context::*;
+pub use direct_context::*;
 pub use surface_metal::*;
+pub use types::*;
